Zotlabs: More type safety for AccessList class.
Add more type declarations to class attributes and functions.

This should ensure that only strings and null values can be assigned to
the various access list members. This is still a bit loose, as we should
probably aim for lists of channel or group id's instead of a generic
type like a string.

I'll leave that for later, though.